Refractor code (#737)
